+++ src/main/MQ2Main.vcxproj
@@ -495,6 +495,7 @@
     <ClInclude Include="..\..\include\mq\api\ActorAPI.h" />
     <ClInclude Include="..\..\include\mq\api\CommandAPI.h" />
     <ClInclude Include="..\..\include\mq\api\DetourAPI.h" />
+    <ClInclude Include="..\..\include\mq\api\GameObject.h" />
     <ClInclude Include="..\..\include\mq\api\Inventory.h" />
     <ClInclude Include="..\..\include\mq\api\Items.h" />
     <ClInclude Include="..\..\include\mq\api\MacroAPI.h" />
@@ -503,6 +504,7 @@
     <ClInclude Include="..\..\include\mq\api\PluginAPI.h" />
     <ClInclude Include="..\..\include\mq\api\RenderDoc.h" />
     <ClInclude Include="..\..\include\mq\api\Spawns.h" />
+    <ClInclude Include="..\..\include\mq\api\SpawnUtils.h" />
     <ClInclude Include="..\..\include\mq\api\Spells.h" />
     <ClInclude Include="..\..\include\mq\api\Textures.h" />
     <ClInclude Include="..\..\include\mq\base\Base.h" />
