+++ src/main/MQ2Main.vcxproj.filters
@@ -583,6 +583,12 @@
     <ClInclude Include="..\..\include\mq\api\Spawns.h">
       <Filter>Header Files\mq\api</Filter>
     </ClInclude>
+    <ClInclude Include="..\..\include\mq\api\GameObject.h">
+      <Filter>Header Files\mq\api</Filter>
+    </ClInclude>
+    <ClInclude Include="..\..\include\mq\api\SpawnUtils.h">
+      <Filter>Header Files\mq\api</Filter>
+    </ClInclude>
     <ClInclude Include="..\..\include\mq\api\MacroAPI.h">
       <Filter>Header Files\mq\api</Filter>
     </ClInclude>
